+++ services/tracker/tests/unit/test_benchmark_service_disconnect.py
@@ -1,4 +1,3 @@
-import json
 import time
 from asyncio import Semaphore
 from contextlib import asynccontextmanager
@@ -173,108 +172,6 @@ class TestBenchmarkServiceDisconnect:
         assert len(sandbox_sources) == 1
         assert isinstance(sandbox_sources[0], SnapshotSource)
         assert sandbox_sources[0].snapshot == "code-migration-eval-container-v8"
-
-    async def test_process_task_injects_tracker_owned_attribution_env(
-        self,
-        contract: AgentContractRequest,
-        database_session: Session,
-        process_benchmark_env: None,
-        monkeypatch: pytest.MonkeyPatch,
-        harness_config: HarnessConfig,
-    ) -> None:
-        contract = contract.model_copy(update={"secrets": {"UNRELATED_SECRET": "secret-name"}})
-        run_starter = RequestIdentity(
-            org=self._test_org,
-            access_key_id="access-key-id",
-            email="starter@example.com",
-            name="Starter User",
-        )
-        start_benchmark_request, task_row, benchmark_id = self._create_task_env(
-            contract,
-            database_session,
-            harness_config,
-            run_starter,
-        )
-        start_benchmark_request = start_benchmark_request.model_copy(
-            update={
-                "benchmark_name": "transient-benchmark-name",
-                "contract": contract.model_copy(update={"name": "transient-agent-name"}),
-            }
-        )
-        captured_env_vars: list[dict[str, str]] = []
-
-        def _mock_resolve_secrets(*_args: Any, **_kwargs: Any) -> dict[str, str]:
-            return {
-                "RUN_ID": "secret-run-id",
-                "TASK_ID": "secret-task-id",
-                "IDENTITY": '{"source":"secret"}',
-                "UNRELATED_SECRET": "secret-value",
-                "MODEL_GATEWAY_URL": "https://gateway.example.test",
-                "MODEL_GATEWAY_API_KEY": "gateway-key",
-            }
-
-        @asynccontextmanager
-        async def _capture_create_sandbox(*_args: Any, env_vars: dict[str, str], **_kwargs: Any):
-            captured_env_vars.append(env_vars)
-            yield SimpleNamespace(id="mock-sandbox-id", name="mock-sandbox-name")
-
-        monkeypatch.setattr(utils_module, "resolve_secrets", _mock_resolve_secrets)
-        monkeypatch.setattr(utils_module, "create_sandbox", _capture_create_sandbox)
-
-        result = await self._run_process_task(start_benchmark_request, task_row, benchmark_id, harness_config)
-
-        assert result == {"task_0": {"status": "success", "score": 1.0}}
-        assert len(captured_env_vars) == 1
-        env_vars = captured_env_vars[0]
-        assert env_vars["RUN_ID"] == str(benchmark_id)
-        assert "QUESTION_ID" not in env_vars
-        assert env_vars["TASK_ID"] == "task_0"
-        assert json.loads(env_vars["IDENTITY"]) == {
-            "benchmark_name": "swebench",
-            "agent_name": contract.name,
-            "email": "starter@example.com",
-        }
-        assert env_vars["UNRELATED_SECRET"] == "secret-value"
-        assert env_vars["MODEL_GATEWAY_URL"] == "https://gateway.example.test"
-        assert env_vars["MODEL_GATEWAY_API_KEY"] == "gateway-key"
-
-    async def test_process_task_omits_identity_email_when_unavailable(
-        self,
-        contract: AgentContractRequest,
-        database_session: Session,
-        process_benchmark_env: None,
-        monkeypatch: pytest.MonkeyPatch,
-        harness_config: HarnessConfig,
-    ) -> None:
-        start_benchmark_request, task_row, benchmark_id = self._create_task_env(
-            contract,
-            database_session,
-            harness_config,
-        )
-        captured_env_vars: list[dict[str, str]] = []
-
-        @asynccontextmanager
-        async def _capture_create_sandbox(*_args: Any, env_vars: dict[str, str], **_kwargs: Any):
-            captured_env_vars.append(env_vars)
-            yield SimpleNamespace(id="mock-sandbox-id", name="mock-sandbox-name")
-
-        def _mock_resolve_no_secrets(*_args: Any, **_kwargs: Any) -> dict[str, str]:
-            return {}
-
-        monkeypatch.setattr(utils_module, "resolve_secrets", _mock_resolve_no_secrets)
-        monkeypatch.setattr(utils_module, "create_sandbox", _capture_create_sandbox)
-
-        result = await self._run_process_task(start_benchmark_request, task_row, benchmark_id, harness_config)
-
-        assert result == {"task_0": {"status": "success", "score": 1.0}}
-        assert len(captured_env_vars) == 1
-        env_vars = captured_env_vars[0]
-        assert json.loads(env_vars["IDENTITY"]) == {
-            "benchmark_name": "swebench",
-            "agent_name": contract.name,
-        }
-        assert "MODEL_GATEWAY_URL" not in env_vars
-        assert "MODEL_GATEWAY_API_KEY" not in env_vars
 
     async def test_invalid_status_produces_human_readable_message(
         self,
